+++ ecchronos-binary/src/pylib/ecchronoslib/types.py
@@ -47,7 +47,6 @@ class RepairJob:
     def __init__(self, data):
         self.keyspace = data["keyspace"] if "keyspace" in data else "<UNKNOWN>"
         self.table = data["table"] if "table" in data else "<UNKNOWN>"
-        self.repair_interval_in_ms = int(data["repairIntervalInMs"] if "repairIntervalInMs" in data else 0)
         self.last_repaired_at_in_ms = int(data["lastRepairedAtInMs"] if "lastRepairedAtInMs" in data else -1)
         self.repaired_ratio = float(data["repairedRatio"] if "repairedRatio" in data else 0)
         self.status = data["status"] if "status" in data else "<UNKNOWN>"
@@ -62,6 +61,7 @@ class RepairJob:
     def get_next_repair(self):
         return datetime.datetime.fromtimestamp(self.next_repair_in_ms / 1000).strftime('%Y-%m-%d %H:%M:%S')
 
+
 class VerboseRepairJob(RepairJob):
     def __init__(self, data):
         RepairJob.__init__(self, data)
@@ -69,3 +69,24 @@ class VerboseRepairJob(RepairJob):
         if "virtualNodeStates" in data:
             for vnode_data in data["virtualNodeStates"]:
                 self.vnode_states.append(VnodeState(vnode_data))
+
+
+class TableConfig:
+    def __init__(self, data):
+        self.keyspace = data["keyspace"] if "keyspace" in data else "<UNKNOWN>"
+        self.table = data["table"] if "table" in data else "<UNKNOWN>"
+        self.repair_interval_in_ms = int(data["repairIntervalInMs"] if "repairIntervalInMs" in data else 0)
+        self.repair_parallelism = data["repairParallelism"] if "repairParallelism" in data else "<UNKNOWN>"
+        self.repair_unwind_ratio = float(data["repairUnwindRatio"] if "repairUnwindRatio" in data else 0)
+        self.repair_warning_time_in_ms = int(data["repairWarningTimeInMs"] if "repairWarningTimeInMs" in data else 0)
+        self.repair_error_time_in_ms = int(data["repairErrorTimeInMs"] if "repairErrorTimeInMs" in data else 0)
+
+    def get_repair_interval(self):
+        return parse_interval(self.repair_interval_in_ms)
+
+    def get_repair_warning_time(self):
+        return parse_interval(self.repair_warning_time_in_ms)
+
+    def get_repair_error_time(self):
+        return parse_interval(self.repair_error_time_in_ms)
+